second commit

+++ __init__.py
@@ -1,3 +1,5 @@
 import os
+import platform
 
-os.system("find . -name '.DS_Store' -type f -delete")
+if platform.system() == "darwin":
+    os.system("find . -name '.DS_Store' -type f -delete")
